requirements.txt
crewai
crewai-tools
python-dotenv
openai
langchain-google-genai
google-generativeai
flask
requests
pydantic
streamlit